jQuery script for searching and playing songs
A get request is made when a user searches, this fetches the html data
and injects it into the container div. Each play button is linked to
youtubeinmp3 to fetch the mp3 of the video link. The mp3 is loaded into
the audio tag and can be controlled by the player div.